# Fix postgres performance regression