--- a/python-sdk/src/astro/databases/base.py
+++ b/python-sdk/src/astro/databases/base.py
@@ -18,8 +18,6 @@ from astro.constants import (
     MergeConflictStrategy,
 )
 from astro.exceptions import DatabaseCustomError, NonExistentTableException
-from astro.files import File, resolve_file_path_pattern
-from astro.files.types import create_file_type
 from astro.files.types.base import FileType as FileTypeConstants
 from astro.settings import LOAD_TABLE_AUTODETECT_ROWS_COUNT, SCHEMA
 from astro.sql.table import BaseTable, Metadata
@@ -28,6 +26,9 @@ from sqlalchemy import column, insert, select
 from sqlalchemy.sql import ClauseElement
 from sqlalchemy.sql.elements import ColumnClause
 from sqlalchemy.sql.schema import Table as SqlaTable
+
+from astro.files import File, resolve_file_path_pattern
+from astro.files.types import create_file_type
 
 
 class BaseDatabase(ABC):
@@ -454,6 +455,13 @@ class BaseDatabase(ABC):
                 if_exists="append",
                 chunk_size=chunk_size,
             )
+
+    @staticmethod
+    def get_dataframe_from_file(file: File):
+        """Get pandas dataframe file
+        :param file: File path and conn_id for object stores
+        """
+        return file.export_to_dataframe()
 
     def load_file_to_table_using_pandas(
         self,