--- a/python-sdk/src/astro/databases/base.py
+++ b/python-sdk/src/astro/databases/base.py
@@ -18,8 +18,6 @@ from astro.constants import (
     MergeConflictStrategy,
 )
 from astro.exceptions import DatabaseCustomError, NonExistentTableException
-from astro.files import File, resolve_file_path_pattern
-from astro.files.types import create_file_type
 from astro.files.types.base import FileType as FileTypeConstants
 from astro.settings import LOAD_TABLE_AUTODETECT_ROWS_COUNT, SCHEMA
 from astro.sql.table import BaseTable, Metadata
@@ -28,6 +26,9 @@ from sqlalchemy import column, insert, select
 from sqlalchemy.sql import ClauseElement
 from sqlalchemy.sql.elements import ColumnClause
 from sqlalchemy.sql.schema import Table as SqlaTable
+
+from astro.files import File, resolve_file_path_pattern
+from astro.files.types import create_file_type
 
 
 class BaseDatabase(ABC):
@@ -458,14 +459,14 @@ class BaseDatabase(ABC):
     @staticmethod
     def get_dataframe_from_file(file: File):
         """
-        Get pandas dataframe file
+        Get pandas dataframe file. We need export_to_dataframe() for Biqqery,Snowflake and Redshift except for Postgres.
+        For postgres we are overriding this method and using export_to_dataframe_via_byte_stream().
+        export_to_dataframe_via_byte_stream copies files in a buffer and then use that buffer to ingest data.
+        With this approach we have significant performance boost for postgres.
 
         :param file: File path and conn_id for object stores
         """
-        # We need export_to_dataframe() for Biqqery, Snowflake and Redshift except for Postgres. For postgres
-        # we are overriding this method and using export_to_dataframe_via_byte_stream(). export_to_dataframe_via_
-        # byte_stream copies files in a buffer and then use that buffer to ingest data. With this approach
-        # we have significant performance boost for postgres.
+
         return file.export_to_dataframe()
 
     def load_file_to_table_using_pandas(