--- a/python-sdk/src/astro/databases/postgres.py
+++ b/python-sdk/src/astro/databases/postgres.py
@@ -13,6 +13,8 @@ from astro.settings import POSTGRES_SCHEMA
 from astro.sql.table import BaseTable, Metadata
 from psycopg2 import sql as postgres_sql
 
+from astro.files import File
+
 DEFAULT_CONN_ID = PostgresHook.default_conn_name
 
 
@@ -186,3 +188,10 @@ class PostgresDatabase(BaseDatabase):
 
         sql = query.as_string(self.hook.get_conn())
         self.run_sql(sql=sql)
+
+    @staticmethod
+    def get_dataframe_from_file(file: File):
+        """Get pandas dataframe file
+        :param file: File path and conn_id for object stores
+        """
+        return file.export_to_dataframe_via_byte_stream()