--- a/python-sdk/src/astro/databases/postgres.py
+++ b/python-sdk/src/astro/databases/postgres.py
@@ -9,11 +9,10 @@ import sqlalchemy
 from airflow.providers.postgres.hooks.postgres import PostgresHook
 from astro.constants import DEFAULT_CHUNK_SIZE, LoadExistStrategy, MergeConflictStrategy
 from astro.databases.base import BaseDatabase
+from astro.files import File
 from astro.settings import POSTGRES_SCHEMA
 from astro.sql.table import BaseTable, Metadata
 from psycopg2 import sql as postgres_sql
-
-from astro.files import File
 
 DEFAULT_CONN_ID = PostgresHook.default_conn_name