--- a/python-sdk/src/astro/files/base.py
+++ b/python-sdk/src/astro/files/base.py
@@ -101,7 +101,7 @@ class File(Dataset):
         structures.
         Due to noted issues with using smart_open with pandas (like
         https://github.com/RaRe-Technologies/smart_open/issues/524), we create a BytesIO or StringIO buffer
-        before exporting to a dataframe. We've found a sizable speed improvement with this optimizat
+        before exporting to a dataframe. We've found a sizable speed improvement with this optimization
 
         :returns: an io object that can be streamed into a dataframe (or other object)
         """