--- a/python-sdk/src/astro/files/base.py
+++ b/python-sdk/src/astro/files/base.py
@@ -116,7 +116,7 @@ class File(Dataset):
         return remote_obj_buffer
 
     def export_to_dataframe_via_byte_stream(self, **kwargs) -> pd.DataFrame:
-        """Read file from all supported location and convert them into dataframes.
+        """Read files from all supported locations and convert them into dataframes.
         Due to noted issues with using smart_open with pandas (like
         https://github.com/RaRe-Technologies/smart_open/issues/524), we create a BytesIO or StringIO buffer
         before exporting to a dataframe. We've found a sizable speed improvement with this optimization.