--- a/python-sdk/tests/databases/test_all_databases.py
+++ b/python-sdk/tests/databases/test_all_databases.py
@@ -5,9 +5,10 @@ from unittest import mock
 import pandas as pd
 import pytest
 from astro.constants import Database
-from astro.files import File
 from astro.settings import SCHEMA
 from astro.sql.table import Metadata, Table
+
+from astro.files import File
 
 CWD = pathlib.Path(__file__).parent
 
@@ -87,7 +88,7 @@ def test_export_to_dataframe_via_byte_stream_is_called_for_postgres(
     database_table_fixture,
 ):
     """Test export_to_dataframe_via_byte_stream() is called in case the db is postgres."""
-    database, table = database_table_fixture
+    database, _ = database_table_fixture
     file = File(str(pathlib.Path(CWD.parent, "data/sample.csv")))
 
     with mock.patch(