--- a/python-sdk/tests/databases/test_all_databases.py
+++ b/python-sdk/tests/databases/test_all_databases.py
@@ -5,9 +5,10 @@ from unittest import mock
 import pandas as pd
 import pytest
 from astro.constants import Database
-from astro.files import File
 from astro.settings import SCHEMA
 from astro.sql.table import Metadata, Table
+
+from astro.files import File
 
 CWD = pathlib.Path(__file__).parent
 
@@ -83,20 +84,19 @@ def test_export_table_to_pandas_dataframe(
     indirect=True,
     ids=["bigquery", "postgres", "snowflake", "sqlite"],
 )
+@mock.patch("astro.files.base.File.export_to_dataframe")
+@mock.patch("astro.files.base.File.export_to_dataframe_via_byte_stream")
 def test_export_to_dataframe_via_byte_stream_is_called_for_postgres(
+    export_to_dataframe_via_byte_stream,
+    export_to_dataframe,
     database_table_fixture,
 ):
     """Test export_to_dataframe_via_byte_stream() is called in case the db is postgres."""
     database, _ = database_table_fixture
     file = File(str(pathlib.Path(CWD.parent, "data/sample.csv")))
 
-    with mock.patch(
-        "astro.files.base.File.export_to_dataframe"
-    ) as export_to_dataframe, mock.patch(
-        "astro.files.base.File.export_to_dataframe_via_byte_stream"
-    ) as export_to_dataframe_via_byte_stream:
-        database.get_dataframe_from_file(file)
-        if database.sql_type == "postgresql":
-            export_to_dataframe_via_byte_stream.assert_called()
-        else:
-            export_to_dataframe.assert_called()
+    database.get_dataframe_from_file(file)
+    if database.sql_type == "postgresql":
+        export_to_dataframe_via_byte_stream.assert_called()
+    else:
+        export_to_dataframe.assert_called()