# Description
## What is the current behavior?
Currently, the performance for loading files in the Postgres database is way too much slower then what it use to be.

OLD Benchmark
GCS to Postgres
Git Sha:  56ac86a1b42fbe2110becd4c226969809fd67147
Run 1
| database   | dataset    | total_time   | memory_rss   | cpu_time_user   | cpu_time_system   | memory_pss   | memory_shared   |
|:-----------|:-----------|:-------------|:-------------|:----------------|:------------------|:-------------|:----------------|
| postgres   | five_gb    | 56.07s       | 45.02MB      | 32.01s          | 14.81s            | 42.36MB      | 7.48MB          |
| postgres   | hundred_kb | 867.03ms     | 32.41MB      | 780.0ms         | 40.0ms            | 29.51MB      | 7.96MB          |
| postgres   | one_gb     | 25.55s       | 37.09MB      | 15.43s          | 8.37s             | 34.13MB      | 7.23MB          |
| postgres   | ten_gb     | 2.71min      | 43.75MB      | 1.38min         | 59.72s            | 40.9MB       | 7.59MB          |
| postgres   | ten_kb     | 888.75ms     | 66.29MB      | 800.0ms         | 80.0ms            | 64.94MB      | 18.96MB         |
| postgres   | ten_mb     | 3.0s         | 55.53MB      | 2.15s           | 220.0ms           | 52.78MB      | 7.77MB          |


Run 2
| database   | dataset    | total_time   | memory_rss   | cpu_time_user   | cpu_time_system   | memory_pss   | memory_shared   |
|:-----------|:-----------|:-------------|:-------------|:----------------|:------------------|:-------------|:----------------|
| postgres   | five_gb    | 51.39s       | 43.38MB      | 32.21s          | 14.75s            | 40.9MB       | 7.32MB          |
| postgres   | hundred_kb | 760.32ms     | 32.46MB      | 780.0ms         | 70.0ms            | 29.75MB      | 7.8MB           |
| postgres   | one_gb     | 23.59s       | 36.65MB      | 15.49s          | 7.7s              | 33.89MB      | 7.22MB          |
| postgres   | ten_gb     | 2.49min      | 43.08MB      | 1.34min         | 59.42s            | 40.55MB      | 7.19MB          |
| postgres   | ten_kb     | 857.16ms     | 61.13MB      | 850.0ms         | 60.0ms            | 59.03MB      | 19.54MB         |
| postgres   | ten_mb     | 2.99s        | 54.93MB      | 2.17s           | 220.0ms           | 52.38MB      | 7.65MB          |


Current main branch:
GCS to Postgres:
Git Sha: 187ad9b
branch: main 
| database   | dataset    | total_time   | memory_rss   | cpu_time_user   | cpu_time_system   |
|:-----------|:-----------|:-------------|:-------------|:----------------|:------------------|
| postgres   | hundred_kb | 500.05ms     | 32.47MB      | 520.0ms         | 30.0ms            |
| postgres   | one_gb     | 6.86min      | 1.24GB       | 3.08min         | 46.24s            |
| postgres   | ten_kb     | 541.13ms     | 63.17MB      | 500.0ms         | 60.0ms            |
| postgres   | ten_mb     | 3.62s        | 50.15MB      | 1.5s            | 120.0ms           |


closes: #876


## What is the new behavior?
There were two bottlenecks found:
1. Reading files in memory and populating them in Postgres - introduced in commit https://github.com/astronomer/astro-sdk/tree/postgres-imp-2
2. The way we were doing dataframe concat was suboptimal.

Branch: postgres_performance_boost

| database   | dataset    | total_time   | memory_rss   | cpu_time_user   | cpu_time_system   |
|:-----------|:-----------|:-------------|:-------------|:----------------|:------------------|
| postgres   | hundred_kb | 905.51ms     | 32.18MB      | 940.0ms         | 70.0ms            |
| postgres   | one_gb     | 1.88min      | 1.08GB       | 1.14min         | 15.41s            |
| postgres   | ten_kb     | 950.24ms     | 69.5MB       | 940.0ms         | 80.0ms            |
| postgres   | ten_mb     | 3.1s         | 54.19MB      | 2.44s           | 190.0ms           |


## Does this introduce a breaking change?
Nope

### Checklist
- [ ] Created tests which fail without the change (if possible)
- [ ] Extended the README / documentation, if necessary